Move connection status notice to chatfooter
