+++ packages/odie-client/src/components/notices/index.ts
@@ -1,0 +1,2 @@
+export { default as useConnectionStatusNotice } from './use-connection-status-notice';
+export { default as useMessageSizeErrorNotice } from './use-message-size-error-notice';
