--- packages/odie-client/src/components/notices/index.tsx
@@ -1,30 +1,0 @@
-import './style.scss';
-import { Icon, close } from '@wordpress/icons';
-import { useOdieAssistantContext } from '../../context';
-import { ConnectionStatus } from '../connection-status';
-
-interface NoticeProps {
-	content?: string | React.ReactNode;
-	onClose?: () => void;
-}
-
-export const Notice: React.FC< NoticeProps > = ( { content, onClose } ) => {
-	return (
-		<div className="odie-notice">
-			{ content && <span>{ content }</span> }
-			<button className="odie-notice__close-button" onClick={ onClose }>
-				<Icon icon={ close } size={ 12 } />
-			</button>
-		</div>
-	);
-};
-
-export const Notices = () => {
-	const { chat } = useOdieAssistantContext();
-
-	return (
-		<div className="odie-notices">
-			{ chat.provider?.startsWith( 'zendesk' ) && <ConnectionStatus /> }
-		</div>
-	);
-};
